Move decorators to parcels and remove wildcard import
