+++ tests-v3/tools/test_helpers.py
@@ -4,7 +4,7 @@ import numpy as np
 import pytest
 
 import parcels.utils._helpers as helpers
-from parcels.utils._decorators import deprecated, deprecated_made_private
+from parcels._decorators import deprecated, deprecated_made_private
 from parcels.utils._helpers import timedelta_to_float
